add CMake and QAM schemes file

+++ src/qam_classes/generator_gauss_noise.cc
@@ -2,11 +2,11 @@
 #include <ctime>
 #include <cmath>
 
-Generator_gauss_noise::Generator_gauss_noise(){ srand(time(0)); }
+Generator_gauss_noise::Generator_gauss_noise(){ srand((unsigned int)time(0)); }
 
 Generator_gauss_noise::Generator_gauss_noise(double noise)
 {
-	srand(time(0));
+	srand((unsigned int)time(0));
 	noise_ = noise;
 }
